Move TaskCompletionSourceEx to Client
