+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/Internal/Threading.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 197d7d8384a6811499a5007c36adc56b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
