+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/Internal/Threading/Tasks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39193f83dd870854282d645ae6325b67
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
