+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/Internal/Threading/Tasks/TaskCompletionSourceEx.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f4d9d79a6f3633849a592bf468988605
+guid: 74bb25518a7ea4d41bd7eeeefd4e09d7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
