--- src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Utils/ReflectionExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e1908e03db2a94e40a051ca5c57759ce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
