--- src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Utils/ReservedWhenAllPromise.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f25f5debe9b9bc2449c5cda7e9cc8994
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
