--- src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Utils/TaskCompletionSourceEx.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f3c7b4e797f7bf8448108f5402481b98
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
