+++ src/MagicOnion.Client/StreamingHubClientBase.cs
@@ -11,6 +11,7 @@ using MagicOnion.Server;
 using System.Buffers;
 using System.Linq;
 using MagicOnion.Client.Internal;
+using MagicOnion.Client.Internal.Threading.Tasks;
 using MagicOnion.Serialization;
 
 namespace MagicOnion.Client
